+++ packages/upload/vite.config.ts
@@ -1,12 +1,55 @@
-/// <reference types="vitest" />
-import { defineConfig } from 'vite'
+import { UserConfig, defineConfig } from 'vite'
 import vue from '@vitejs/plugin-vue'
+import { resolve } from 'path'
+import dts from 'vite-plugin-dts';
 
 // https://vitejs.dev/config/
 export default defineConfig({
-  test: {
-    globals: true,
-    environment: "jsdom",
-  },
-  plugins: [vue()],
-})
+  plugins: [
+    dts({
+      outDir: ['./dts'],
+      include: ['./src/'],
+    }),
+    vue()
+  ],
+  base: './',
+  build: {
+    target: 'modules',
+    //打包文件目录
+    outDir: 'es',
+    //压缩
+    minify: true,
+    //css分离
+    //cssCodeSplit: true,
+    rollupOptions: {
+      //忽略打包vue、element-plus
+      external: ['vue'],
+      input: ['./src/main.ts'],
+      output: [
+        {
+          format: 'es',
+          //不用打包成.es.js,这里我们想把它打包成.js
+          entryFileNames: '[name].js',
+          //让打包目录和我们目录对应
+          preserveModules: true,
+          exports: 'named',
+          //配置打包根目录
+          dir: resolve(__dirname, './ui/es'),
+        },
+        {
+          format: 'cjs',
+          entryFileNames: '[name].js',
+          //让打包目录和我们目录对应
+          preserveModules: true,
+          exports: 'named',
+          //配置打包根目录
+          dir: resolve(__dirname, './ui/lib'),
+        },
+      ],
+    },
+    lib: {
+      entry: './index.js',
+      formats: ['es', 'cjs'],
+    },
+  }
+}) as UserConfig