+++ example/CHANGELOG.md
@@ -1,5 +1,40 @@
 # example
 
+## 1.2.34
+
+### Patch Changes
+
+- Updated dependencies
+  - @tempoi/search-form@1.2.34
+
+## 1.2.33
+
+### Patch Changes
+
+- Updated dependencies
+  - @tempoi/search-form@1.2.33
+
+## 1.2.32
+
+### Patch Changes
+
+- Updated dependencies
+  - @tempoi/search-form@1.2.32
+
+## 1.2.31
+
+### Patch Changes
+
+- Updated dependencies
+  - @tempoi/search-form@1.2.31
+
+## 1.2.30
+
+### Patch Changes
+
+- Updated dependencies
+  - @tempoi/search-form@1.2.30
+
 ## 1.2.29
 
 ### Patch Changes